+++ concurrency/v3/websiteChecker_test.go
@@ -1,6 +1,9 @@
 package concurrency
 
-import "testing"
+import (
+	"testing"
+	"time"
+)
 
 func TestWebsiteChecker(t *testing.T) {
 	websites := []string{
@@ -11,11 +14,11 @@ func TestWebsiteChecker(t *testing.T) {
 
 	expectedResults := []bool{
 		true,
-		true,
 		false,
+		true,
 	}
 
-	actualResults := WebsiteChecker(websites)
+	actualResults := WebsiteChecker(fakeIsWebsiteOK, websites)
 
 	want := len(websites)
 	got := len(actualResults)
@@ -28,11 +31,34 @@ func TestWebsiteChecker(t *testing.T) {
 	}
 }
 
+func BenchmarkWebsiteCheckerWithManyURLs(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		websites := make([]string, 100)
+		for index, _ := range websites {
+			websites[index] = "http://google.co.uk"
+		}
+
+		WebsiteChecker(slowIsWebsiteOK, websites)
+	}
+}
+
 func sameResults(as, bs []bool) bool {
 	for index, a := range as {
 		if a != bs[index] {
 			return false
 		}
+	}
+	return true
+}
+
+func slowIsWebsiteOK(_ string) bool {
+	time.Sleep(20 * time.Millisecond)
+	return true
+}
+
+func fakeIsWebsiteOK(url string) bool {
+	if url == "http://blog.gypsydave5.com" {
+		return false
 	}
 	return true
 }
