--- concurrency/v3_di/refactor/websiteChecker.go
@@ -1,13 +1,0 @@
-package concurrency
-
-type testURL func(string) bool
-
-func websiteChecker(isOK testURL, urls []string) []bool {
-	results := make([]bool, len(urls))
-
-	for index, url := range urls {
-		results[index] = isOK(url)
-	}
-
-	return results
-}
