--- concurrency/v3_di/test_passes/isWebsiteOK.go
@@ -1,19 +1,0 @@
-package concurrency
-
-import (
-	"net/http"
-)
-
-// IsWebsiteOK returns true if the URL returns a 200 status code, false otherwise
-func IsWebsiteOK(url string) bool {
-	response, err := http.Head(url)
-	if err != nil {
-		return false
-	}
-
-	if response.StatusCode != http.StatusOK {
-		return false
-	}
-
-	return true
-}
